requirements.txt
lxml
python-edgar
pandas
pyyaml